gitserver: use migration cursor to decide which hashing algorithm to use during gitserver address determination
The cursor is a name of last repository which was processed (i.e. migrated to new gitserver instance in most cases) during the process of hashing algorithm change. Any repo with a lexicographically smaller or equal name than cursor is located on a gitserver instance which hostname is resolved using rendezvous hashing.