Avoid superfluous query parsing
I noticed that the query input was parsed twice on every keypress. Turns
out that the `onSubmit` function changes because the value changes. This
caused the extensions to be recreated.

Instead of making an extension directly from the submit handler, the
handler is now provided as a configuration value to a state field via an
effect. So while that will still cause the editor state to update
multiple times, at least the query (and therefore the syntax highlights)
won't be computed multiple times for the same input.