search: Add possibilty to use CodeMirror for the search query input (#32446)
This commit introduces CodeMirror for the search query input, which is gated behind the new `editor` experimental feature flag. `LazyMonacoQueryInput` uses that value to determine which editor to load. If set to `codemirror6`, CodeMirror will be used for the search query input. For more information about this change, please see the PR description: https://github.com/sourcegraph/sourcegraph/pull/32446 .