add stream tracing (#32799)
This adds stream tracing to all our jobs. For each event sent down
the stream, we log the number of results in the event. This is useful
for debugging latency and visualizing the frequency of events across the
job tree. Additionally, we log the total number of events streamed by
each job.