store: remove dependency on database.DB (#32883)
We recently introduced the need to have a database.DB to speak to
gitserver. In that commit we threaded through the DB from searcher.
However, we did not need to do that since we already have the DB via the
closure.

On a higher level: The intention of the FetchTar function is to inject
the minimal interface we need to describe what we want. DB does not
change across requests, so we can remove it.

Note: There is also FilterTar which in theory doesn't need DB as the
input. However, the way the function is currently defined it does since
it directly uses the gitserver client.

Test Plan: CI