remove unused exported functions
I started by just removing IsStreamingCompatible since I recently
removed the last references to it, but noticed there were a significant
number of dead exported functions in this package.

I'm happy to resurrect any if they are going to be used, but I think as
part of our backend cleanup, it will be helpful to have packages that
are as lightweight as possible.