[cloud] Case insensitive comparison of emails for invitations
Previously, there was a case sensitive string comparison of the
email on the invitation with the emails configured on the user account.
But email is case insensitive in it's nature, meaning that people
are likely to fill their email as John.Doe@foo.bar instead of john.doe@foo.bar

Doing a case insensitive comparison fixes the problem.

Related:
https://sourcegraph.atlassian.net/browse/CLOUD-337