store: move back into searcher/search (#32957)
store used to live in searcher and mostly be not exported. It was
exported in 82e525866ecf (April 2019) for our replacer service I
believe. However, for a while now the only importer of it has been
searcher. So lets move it back in. This gives a few opportunities to
simplify.

Test Plan: CI