sg: don't magically source .bashrc for keegan
I think we do this because we mutate .bashrc in the sg setup and we try
to validate if stuff is working after mutation. However, I don't use sg
setup and I rely direnv to setup a special environment just for
Sourcegraph.

This commit exists as a talking point for the proper solution. Although
I don't mind landing it either :)

Test Plan: sg works on my linux machine.