use match count, not result count in slack notification
Previously, we used result count in the title, but match count in the "n
more matches" trailer. This is confusing because you might get a message
that says "3 new matches" but then a trailer that says "5 more matches",
which doesn't make any sense. This updates it to use match counts
instead of result counts.