gitserver: add heuristics to skip maintenance if possbile
Running git-repack on every cleanup is very expensive. This adds
heuristics similar to those used by git and GiLlab's gitaly.  The goal
is to skip the maintenance job most of the time.

There is a caveat for repos with a lot of unreachable loose objects:  If
a repo has enough loose objects, we trigger a repack. If a lot of those
loose objects are unreachable, they will remain loose after the repack.
This means we run repack every time until those unreachable
objects are eventually pruned. We could reduce this possibility by
reducing the time we keep the unreachable objects (currently 2 weeks).