unpack: Implement Zip archive unpacking
This commit implements Zip archive unpacking, leveraging the existing
test suite for all unpackers to cover its functionality.

The auxiliary `fileInfo` type changed to implement symlinks with their
destination as the contents of the file, like it's done in Linux and in
Zip archives in order to support symlinks in Zip.

Tar symlinks remain header only files in the tar archive.