extsvc: Introduce gomodproxy.Client
This commit introduces a client of Go module proxies, with two methods:
`GetVersion` and `GetZip`.

The client can be configured with multiple proxy URLs, similarly to how
the environment variable `GOPROXY` in a Go workspace functions, except
that `direct` mode is not supported (which would mean talking the Git
protocol instead).

This allows a customer to configure a private module proxy for private
modules and a public one for everything else.