dev/dn: add ServicesPerPullRequest and ManifestRevision (#33119)
Adds ServicesPerPullRequest, which is an accurate representation of exactly which pull requests are associated with each service in the deployment because each service might be deployed with a different source diff. This is important for notifications, tracing, etc, and the information is currently discarded by the report's PRs and services summary.

Also adds ManifestRevision for identifying the deployed revision.