search: Remove unindexed search limiting
The existing unindexed search limiting to 200 repo revisions was
introduced in #11546 before we had paginated repository resolution,
which has a page size of 500 and naturally limits the number of
concurrent unindexed searches happening per-request.

This is more than double of the previous limit, but nowhere close the
originally reported problematic number of 20k.

Fixes #33214