New reference panel: sort search-based results by proximity (#33223)
This ports the jaccard-index-based sorting from code-intel-extensions to
the new reference panel and adjusts the types for it to work properly.

I also need to double-check whether the old code isn't broken in
code-intel-extensions, because it uses the `url.hash` to get to the
"file path" and our current URL schema doesn't have that in the hash.
But maybe the extension-api uses different schemes. Will check.