gitserver: Skip remote urls for private cloud default repos
We won't be able to use a remote URL for a private repo linked to a
cloud default external service, so we should skip it. This can happen if
a repo moves from being public to private while belonging to both a
cloud default external service and another external service with a token
that has access to the private repo.