Simplify use of GQL fragments to avoid warnings
This gets rid of some of the warnings when booting up VS Code, but I
can't get rid of the last tone, which is due to the `codeIntelFragments`
being injected.

Seems like eslint/TS/whatever only understand it once the file is
changed & saved.

This still helps with _some_ of it.