search-blitz: extra queries for dogfood (#33505)
We want to run search blitz on dogfood. However, we need extra queries
to target gigarepo which is only available on dogfood. This commit adds
logic to take the existing queries and add extra queries based on the
environment variable SEARCH_BLITZ_ENV.

Test Plan: CI and manual testing on dogfood.