search: Fix user settings overwriting settings from URL
The logic to load search parameters from user settings only accounted
for search parameters passed as actual URL parameters, not as part of
the search query.

This PR changes the approach and moves the responsiblity of correctly
applying the settings from the user settings loading code to the query
state store.