gitserver: eventually re-clone after failed sg maintenance (#33432)
It is possible that sg maintenance fails on a corrupt repo. If this is
the case, we keep running sg maintenance periodically until the repo is
eventually fixed (either manually or due to a re-clone).

Git's gc deals with this by placing a gc.log file in the git dir. This
change follows the same logic.

A failed sg maintenance run will place a log file in the git directory.
Subsequent sg maintenance runs are skipped unless the log file is old.
A successful run removes the log file.

If `maybeReclone` encounters the log file, it will re-clone the repo.