ci: run go steps for any change in go sub dirs
Non go files can also affect the outcome of go build (embed) and go test
(testdata). We hardcode a list of directories which mostly contain go
code and run go steps if a change includes them.

A recent example of this is where I only modified a file we go:embed
which caused the tests to fail. However, CI did not run the tests on the
PR leading to a failure on main.

Test Plan: go test