repo-updater: Sync rate limiter before service
If it is a newly added service then syncing the service first means we
won't have a rate limiter in place yet and we may fall back to using an
infinite limiter.

In theory the rate limit sync should replace it shortly afterwards but
we're seeing cases of infinite limiters in production and this is one
area that could be the culprit.