gitserver: refactor TestCleanupOldLocks (#33651)
The original tests was too verbose as it checked for files that were
unrelated to the test's objective. This meant we had to update the test
even for unrelated changes, like adding a sgm.log file for failed sg
maintenance runs.

We don't check for repos-stats.json anymore, because this is already
covered in `TestCleanup_computeStats`.