frontend: Replace use of errors.Cause with errors.Is.
Turns out, errors.Unwrap is not a left-identity to errors.Wrap.
Moreover, errors.UnwrapAll can return a more 'bare' error than
the original error created by errors.New. This means that errors.Cause
would have 0 stack frames; comparing that to a result from errors.New
(with 1 stack frame) would fail.
Thanks to this counter-intuitive behavior, switching over the result
of errors.Cause is (ahem) error-prone; it MUST NOT be compared directly
with the result of errors.New.

So let's not use errors.Cause + switch; let's use errors.Is instead.

Fixes issue 33697; we had code paths for making sure that we showed
the unhighlighted file on HSS timeouts, but that code path was not
triggered because we weren't checking the error's identity properly.