searcher: hybrid search with zoekt prototype
This lays the ground work to consult the zoekt index for unindexed
searches. We search the zoekt index for files that are the same for our
commit, and then fallback to searching just the changed paths via
unindexed search. This results in us fetching much less data from
gitserver.

There is some changes to the frontend code to wire through a feature
flag. This code is behind the feature flag "search-hybrid" to make it
safer to merge and experiment.

Test Plan: go test ./cmd/searcher/... manualling searching the megarepo
locally with the feature flag enabled.