Search backend: reorganize job packages (#33741)
This reorganizes the job packages to simplify some package dependencies. The main issue is that the internal/search/job package effectively could not be imported by any other search package because it imported many search packages so it could have a list of allJobs. Importing it would cause circular dependencies.