[cloud][CLOUD-356] Change default invite quota for users
Right now the default quota is 15. However we expect cloud open beta
customers to be able to invite ~100 developers in their organization.

Changing the default quota to 100 to adhere to this expected size.
Otherwise we would need more than 1 person to invite the organization
members (or change the quota in the DB directly on-demand), which is not
desired.