Code monitors: simplify last searched (#33732)
This removes the concept of an "args hash" from repo-aware code monitors. It restricts the types of queries that can be run by code monitors to queries that generate a single commit search job. This way, we only need to know which commits were last searched by the code monitor.