github: Don't capture pointer in collector (#33892)
We're seeing metrics being collected with names of non site owned
external services, this should not be the case as we only set up
collectors for site owned external services. The only way the wrong name
could be coming through is if the pointer to the service that is
captured by the collector closure is changed at some point. It's not
clear where this could happen, but by passing the name through as a
value we avoid this possibility.