internal/authz: Add integration test for full user + repo sync
This commit adds useful context in the form of TODO comments to
explain why we expect the test to fail. The part expected to fail is
commented out and future work will take advantage of this test which
is when the commented bits will be uncommented and the test will shine
for its true potential!