repo-updater: Include owners when syncing external service (#33995)
We were sending the owners in the sync request but not including them
when creating the external services struct. This meant that it would
appear as a site owned external service.

At the very least, this us causing us to log metrics against the wrong
external service and it may be causing bugs in other areas of the code
base.