gitserver: support suffix .git for go modules  (#33942)
With this change we don't remove the suffix ".git" if the code host is a go module proxy. Go supports module names with suffix ".git", see for example git.apache.org/thrift.git. So far, we removed the suffix for all code hosts during normalization, which meant we weren't able to clone thrift.git.