gitserver: document Addrs
This was missing the documentation that was originally present when
Addrs was a function field on the gitserver client struct.

Note: I checked the other methods which are not documented, and the
signature for those is so obvious that I don't think documentation
helps.

Test Plan: n/a internal documentation change.