detach computeExcluded from repo resolver
This detaches the `Excluded` method from `repos.Resolver` and makes it a
standalone, private function. Computing excluded repos requires a
slightly different set of information than actually resolving repos.
Specifically, whether `fork` and `archived` were explicitly set. This
makes it possible to only pass that information to the function that
needs it.

The goal here is to minimize passing informaiton into scopes where it
isn't used.