Search backend: remove query from RepoOptions (#34138)
We currently hold a reference to the query in RepoOptions so we can
determine whether fork and archived were set explicitly by the user, or
whether we used defaults for those values. However, since we only need
to know whether they were set explicitly, we can store that instead and
reduce the scope of RepoOptions.