internal/extsvc/github: Fix panic in fetchGitHubVersion (#34172)
In a previous commit (by yours truly), we updated the method to return
a named argument and unintentionally introduced a potential panic.

The call to `semver.NewVersion(v)` overwrites the initialized value of
version at the start of the method. In the case that
`semver.NewVersion` returns an error, it will return `(nil, err)`
leading to `version` being overwritten as `nil`. And consequently this
method no longer returns `allMatchingSemver` by default.

We have a failing test in #34004 which showed up after moving it to a
different file. Note sure why this only showed up now. Creating a
separate PR to keep the scope of the large PR limited, and to preserve
the context of this fix independently.