Search backend: remove passed-in RepoOptions from Paginate signature (#34156)
We've got some duplication in the API for repo search since options can
either be passed in through Paginate or passed in as a field of the
resolver. This removes the less-used first option. This is not an
opinion on which way is better, but rather just a "make it work one way"
thing.