sg: fix panic when docker.GetCredentialsFromStore errors
Previously we always set dockerCredentials from the store. However, if
that failed, that means dockerCredentials would become nil. We then
wrote to it. This commit instead uses an intermediate variable.

Note: this may have worked in the past and regressed due to the docker
dependency being updated.

Test Plan: I don't get a panic when releasing 3.39