[web][CLOUD-265] Fix focus order for `Forgot password?` link
Previously, the link was second in tab focused items on the page.
This was inconvenient, because after typing in username, hitting
tab would focus on the Forgot password link instead of the
password input.

Now the order of the focus was reversed so that password input
is focused before the Forgot password link. Which allows
to fill the login form by hitting TAB only once.

Related:
https://sourcegraph.atlassian.net/browse/CLOUD-265