dev: go-lint only looks for go.mod files (#34325)
We had "-type file" which broke since this doesn't work on linux. This
was removed, but now we get a test directory called go.mod matching
which is incorrect. This adds back the find condition, but correctly
works on both mac and linux.

  $ diff <(find . -name go.mod -type f) <(find . -name go.mod)
  1a2
  > ./internal/codeintel/dependencies/internal/lockfiles/testdata/parse/go.mod

Test Plan: ran the find incantation on both linux and mac. Both agreed
and found all go.mod files.