gitserver experiment: add shortcut for getting gitserver address during SyncRepoState
gitserver.AddrForRepo call was substituted with this private function call to avoid possible querying of the DB during the experiment. SyncRepoState fetches addresses for all gitserver_repos table rows and this will cause extra unnecessary load on DB and spoil metrics data.
This commit will be reverted as soon as the experiment ends.