notebooks: Refactor column layout (#34273)
While working on the notepad CTA (which will be a separate PR), I noticed some minor issues with current layout of the notebook page. In particular:

- Gap/outline overflow appearing with small window height (at least in Firefox)
- Seemingly erratic scroll behavior when scrolling the outline (at least in Firefox)
- Incorrect position of tooltip text for very long headings
- Header and action alignment on small screens.

Instead of using flex box with three columns and making one column sticky (which I assume was done to (1) center the content within the window and (2) have the scrollbar for the main content appear on the right edge of the window), the new approach uses two columns which allows it to make them scrollable independently.

I tried to preserve the current look and feel as much as possible. In particular:

- Fixed outline heading and scrollable headings list
- Scrollbar for main content is positioned at the right edge of the window