syntax-highligher: tiny refactors as we went along
Noah and I were reading through the code and while trying to understand
how it works we played around with some things and tried to refactor
them or move them around.

This is the result. Nothing here is "I need to get this merged, asap!",
it's all "eh, if you like it, cool, if not, I'll close this PR".