Search backend: do not apply `DNF` to `or` expressions (#34382)
This modifies the DNF step to only distribute or expressions that contain parameters. This allows us to maintain the tree structure of patterns in basic queries so that we can generate optimized or queries for our backends. The Zoekt backend already takes advantage of this, and the commit search backend will be able to as well. Comments inline.