add searchblitz for a simple, literal AND query (#34486)
Our searchblitz query exercising `and` queries right now is a fairly
complex query with 4 terms and somewhat expensive regexes. To reflect
the simplest use of `and` to better measure the base overhead of an and
query, this adds the `and_literal_simple` query. This uses the same
query string as `or_literal_simple` so they can be directly compared.