all: use alpine-3.14
We have a 3.14 image available, but when we introduced it we ran into
problems using it. This is an attempt to update all images to use the
newer version of alpine. If anything breaks in CI, I will revert the
specific image which is breaking.

  ruplacer --go \
    sourcegraph/alpine-3.12:142406_2022-04-14_8836ac3499f4@sha256:4681a48d1fb9a73fef1b540c08b3411f797351bbeda749f5dca21213a1e71526 \
    sourcegraph/alpine-3.14:142406_2022-04-14_8836ac3499f4@sha256:2a2d1cbaec78882661fe1aa5b0a4af0c23a37be2ea9ff8aadc2da5b80852c233

Test Plan: main dry run on CI to test docker image building and
integration testing.