Fix incorrect use of errors.Append masking errors
It's possible to pass only a single argument to errors.Append which has
the effect of not appending the errors and only returning the last one.

It's really easy to miss because the API makes it look like you're
calling a Append method on a receiver, which would imply it has a
side-effect. But it's not a receiver it's a package, so it just silently
discard errors.