monitoring: Rename core-application to repo-management and refactor (#30026)
In this commit we introduce a struct to store the owner name, team
name and the slug since now all the teams have a custom format. As a
result, the existing switch-case in `toMarkdown` can be replaced in
exchange of a better data struct-ure (pun intended).

Also, we rename core-application to repo-management. This is the same
team in practice but with an outdated name. There is no change of 
ownership of any alerts in this commit.