Structural search: remove always-false useFullDeadline in zoektSearch (#34588)
This removes the useFullDeadline parameter from zoektSearch, which was
always false. There is a TODO that was removed as part of this to handle
this better, but I think the real solution is just to stream results
into from zoekt and into comby, so I think it's fine to simplify and
leave it as it's been working for the last while.