Search backend: remove Dnf and friends (#34615)
We replaced the functionality of Dnf with BuildPlan, which maintains
tree structure of pattern nodes. I hadn't yet removed the
now-unnecessary Dnf code because I wanted it to be easy to revert, but
it's been out there for a bit with no issues. This replaces uses of Dnf
with BuildPlan, and removes it and associated code. I think there is
more code to simplify here, but I'll leave that for a followup PR.