observation: disable trace logs in dev (#34703)
Explicit logging of trace logs was added because Datadog does not support OT/OTel trace logs/events at INFO level because we do not want these logs to be lost if we have no other way of seeing trace log events. This PR disables it for dev environments, and optionally for other environments, because it can get a bit noisy.