Fix link to docs on empty-auto-indexing page
I ran into this when recording the demo for auto-indexing yesterday. I
was surprised that the link didn't lead me anywhere close to
*auto*-indexing, but just to indexing a Go repository. I suspect it was
once a placeholder?

This now links to the "enable auto-indexing" page from which users can
either directly go to enable it or read up on what it is.