wildcard: Fix missing caret in textarea on Firefox
The Firefox specific styles cause the caret to be transparent in
textareas. The comment seems to suggest that this isn't actually
intended to be used with textareas, but maybe I'm not understanding it
properly. I also inspected our custom and native select components and
those styles don't even seem to be applied to them.