codemirror: Fix initial pattern type not being recognized
The root issue was that the extensions get recreated due to changes
higher up in the component tree. However, because the pattern type is
set via a transaction the current value was lost and the extension
default back to `literal`.

Order of events:

1. Editor gets initialized
    1. Pattern type defaults to `literal`
2. Transaction gets triggered (via hook) to set pattern type to `regex`
3. Editor gets initialized`
    1. Pattern type defaults to `literal`

The hook that sets the pattern type via a transaction is *not* triggered
again because its depdencies (editor instance, pattern type) didn't
change.

The fix is to pass the current pattern type when initializing the
extension.

I also looked over the code to see if there are other extension that
would exhibit the same problem, but this one seems to be the only one.