ci: invalidate asdf cache before using it
There can be an existing .asdf directory when asdf-install.sh is run.
For example in the backcompat tests we checkout an older commit then run
asdf-install. The mixing of the directory leads to undefined behaviour.
In particular for the go1.18 upgrade PR when running tests against the
older commit we keep using the newer version of go. In this particular
case this leads to the tests failing due to new lint errors in go1.18.

Test Plan: main dry run on CI